-updated auth resolvers to use tokens
